+++ RepairsApi/V2/Email/GovUKNotifyService.cs
@@ -2,6 +2,7 @@ using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
 using Notify.Client;
 using Notify.Exceptions;
+using Notify.Interfaces;
 using System.Threading.Tasks;
 
 namespace RepairsApi.V2.Email
@@ -16,21 +17,23 @@ namespace RepairsApi.V2.Email
     {
         private readonly NotifyOptions _options;
         private readonly ILogger _logger;
+        private readonly IAsyncNotificationClient _notificationClientAsync;
 
-        public GovUKNotifyService(IOptions<NotifyOptions> options, ILogger<GovUKNotifyService> logger)
+        public GovUKNotifyService(IOptions<NotifyOptions> options,
+            ILogger<GovUKNotifyService> logger,
+            IAsyncNotificationClient notificationClientAsync)
         {
             _options = options.Value;
             _logger = logger;
+            _notificationClientAsync = notificationClientAsync;
         }
 
         public Task SendMailAsync<TRequest>(TRequest request)
             where TRequest : EmailRequest
         {
-            var notificationClient = new NotificationClient(_options.ApiKey);
-
             try
             {
-                return notificationClient.SendEmailAsync(request.Address, ResolveTemplateId<TRequest>(), request);
+                return _notificationClientAsync.SendEmailAsync(request.Address, ResolveTemplateId<TRequest>(), request);
             }
             catch (NotifyClientException ex)
             {
