+++ RepairsApi/V2/Email/HighCostVariationCreatedEmail.cs
@@ -1,0 +1,11 @@
+﻿namespace RepairsApi.V2.Email
+{
+    internal class HighCostVariationCreatedEmail : EmailRequest
+    {
+        public HighCostVariationCreatedEmail(string address, int id)
+            : base(address)
+        {
+            Set("workOrderId", id);
+        }
+    }
+}
