+++ RepairsApi/V2/Email/VariationApprovedEmail.cs
@@ -1,0 +1,11 @@
+﻿namespace RepairsApi.V2.Email
+{
+    internal class VariationApprovedEmail : EmailRequest
+    {
+        public VariationApprovedEmail(string address, int id)
+            : base(address)
+        {
+            Set("workOrderId", id);
+        }
+    }
+}
