+++ RepairsApi/V2/Email/VariationRejectedEmail.cs
@@ -1,0 +1,11 @@
+﻿namespace RepairsApi.V2.Email
+{
+    internal class VariationRejectedEmail : EmailRequest
+    {
+        public VariationRejectedEmail(string address, int id)
+            : base(address)
+        {
+            Set("workOrderId", id);
+        }
+    }
+}
